feat: token price diff (#25)
* feat: increase PriceGraph max value

Increasing max value lowers the graph and prevents
it from overlaping with balance.

* refactor: usePriceHistory

Refactor usePriceHistory to make it able to be
used as a general function for fetching prices.

* feat: token price difference

For now it's just a bare working version. I wanted
to write tests but it's late and I'm not in the
mood but I'll do so no worries 'bout that.